modify watch.sh to be compatible with Task
it accepts an optional `task` argument which makes it run task commands instead of make commands (the default if no argument is supplied).

since I'm on Windows I had to add a parameter to .gitattributes so I could save watch.sh with LF line endings, otherwise it would fail to be executed in Linux.